requirements.txt
fastapi
uvicorn[standard]
gspread
pandas
python-dotenv
python-jose[cryptography]
pytz
